PC twin: the whole hal.h on desktop SDL2 (roadmap: desktop HAL)
cd sdk/<game> && make -f ../pc/pc.mk -> <game>-pc, a native binary of
the unmodified game: 320x240 palettized window (3x), keyboard as the
pad (arrows/Z/X/A/S/Q/W/TAB/ENTER, ESC quits), 48 kHz audio with
console-FIFO-style backpressure, pak from ./game.pak or $RVSTACK_PAK,
saves as ./<name>.sav. Game logic now iterates at PC speed under a
debugger; the Pocket is only needed for hardware truth. Pattern
validated by openfpgaSDK's PC backend; implementation is ours over
hal.h so games stay single-source. pak_file_t/save_file_t .base
widened to uintptr_t — bit-identical ABI on rv32, correct on x86-64.
Console savetest scenario re-run: PASSES.

+++ soc/hal/hal.h
@@ -159,7 +159,9 @@ int       audio_stream_write(const int16_t *pcm, int nframes);
 // NOTE: usable size = file size - 2 (APF EOF-read wedge bug) — pad pak files.
 // name is ignored for now (one slot); whence: 0=SET, 1=CUR, 2=END.
 
-typedef struct { uint32_t base; uint32_t size; uint32_t pos; } pak_file_t;
+// base is uintptr_t so the PC twin (64-bit) shares this header; on the
+// console (rv32) it is the same 32-bit word as before — ABI unchanged.
+typedef struct { uintptr_t base; uint32_t size; uint32_t pos; } pak_file_t;
 
 int       pak_open(const char *name, pak_file_t *out);   // <0: none/failed [BUILT]
 int       pak_read(pak_file_t *f, void *dst, int nbytes);             // [BUILT]
@@ -184,9 +186,9 @@ void      pak_run_game(const pak_file_t *g);                          // [BUILT]
 // ============================================================================
 
 typedef struct {
-	uint32_t base;                  // your save data, size bytes of DRAM
-	uint32_t size;                  // capacity (request rounded up to 4)
-	char     _path[64];             // internal: SD path bound at open
+	uintptr_t base;                 // your save data, size bytes of DRAM
+	uint32_t  size;                 // capacity (request rounded up to 4)
+	char      _path[64];            // internal: entry name / PC file path
 } save_file_t;
 
 // Open (create if missing) this game's save file. name = a short identifier
